Streaming `Message` documentation and move internal message classes
